chore: moved workflow tests and made them concurrent
